rename folder, remove unused typeguard
