+++ src/components/fields/schemaFields/widgets/cssClassWidgets/CssSpacingWidget.module.scss
@@ -25,3 +25,7 @@
 .spacingControlSelect {
   width: 10rem;
 }
+
+.expandButton {
+  width: 5rem;
+}
